Expose nav mesh and baseplayer functions for non-TF mods
