+++ src/game/server/tf/bot/tf_bot.h
@@ -293,7 +293,7 @@ public:
 
 	void SetHomeArea( CTFNavArea *area );
 	CTFNavArea *GetHomeArea( void ) const;
-	void ScriptSetHomeArea( HSCRIPT hScript ) { this->SetHomeArea( ToNavArea( hScript ) ); }
+	void ScriptSetHomeArea( HSCRIPT hScript ) { this->SetHomeArea( ToTFNavArea( hScript ) ); }
 	HSCRIPT ScriptGetHomeArea( void ) { return ToHScript( this->GetHomeArea() ); }
 
 	CObjectSentrygun *GetEnemySentry( void ) const;			// if we've been attacked/killed by an enemy sentry, this will return it, otherwise NULL
