+++ src/game/server/tf/nav_mesh/tf_nav_area.h
@@ -63,6 +63,8 @@ class CTFNavArea : public CNavArea
 public:
 	DECLARE_CLASS( CTFNavArea, CNavArea );
 
+	DECLARE_ENT_SCRIPTDESC();
+
 	CTFNavArea( void );
 	~CTFNavArea( void );
 
@@ -129,50 +131,6 @@ public:
 	// Distance for MvM bomb delivery
 	float GetTravelDistanceToBombTarget( void ) const;
 
-	//- Script access to nav functions ------------------------------------------------------------------
-	DECLARE_ENT_SCRIPTDESC();
-	HSCRIPT GetScriptInstance();
-	bool IsBottleneck( void ) const;
-	Vector FindRandomSpot( void ) const;						// return a random spot in this area
-	void OnDoorCreated( CBaseEntity *door );					// invoked when a door is created
-	CBaseEntity *GetDoor( void ) const;							// return a door contained in this area
-	
-	int ScriptGetID( void ) const { return (int)GetID(); }
-	void ScriptGetAdjacentAreas( int dir, HSCRIPT hTable );
-	HSCRIPT ScriptGetAdjacentArea( int dir, int i );
-	HSCRIPT ScriptGetRandomAdjacentArea( int dir );
-	void ScriptGetIncomingConnections( int dir, HSCRIPT hTable );
-	void ScriptAddIncomingConnection( HSCRIPT hSource, int incomingEdgeDir );
-	void ScriptConnectToArea( HSCRIPT hArea, int dir );
-	void ScriptDisconnectArea( HSCRIPT hArea );
-	bool ScriptIsConnectedArea( HSCRIPT hArea, int dir );
-	Vector ScriptGetCorner( int corner ) const { return GetCorner( (NavCornerType)corner ); }
-	void ScriptMarkAsBlocked( int teamID );
-	int ScriptGetAdjacentCount( int dir ) const	{ return GetAdjacentCount( (NavDirType)dir ); }
-	const char* ScriptGetPlaceName();
-	void ScriptSetPlaceName( const char* pszName );
-	int ScriptComputeDirection( const Vector &point ) const;
-	int ScriptGetPlayerCount( int teamID ) const { return GetPlayerCount( teamID ); }
-	bool ScriptIsOverlapping( HSCRIPT hArea ) const;
-	bool ScriptIsOverlappingOrigin( const Vector &pos, float tolerance ) const { return IsOverlapping( pos, tolerance ); }
-	bool ScriptIsEdge( int dir ) const { return IsEdge( (NavDirType) dir ); }
-	bool ScriptContains( HSCRIPT hArea ) const;
-	bool ScriptContainsOrigin( const Vector &pos ) const { return Contains( pos ); }
-	float ScriptComputeGroundHeightChange( HSCRIPT hArea );
-	HSCRIPT ScriptGetParent( void );
-	int ScriptGetParentHow( void ) const { return GetParentHow(); }
-	void ScriptUnblockArea( void );
-	bool ScriptIsVisible( const Vector &eye ) const	{ return IsVisible( eye ); }
-	float ScriptGetZ( const Vector &pos ) const	{ return GetZ( pos ); }
-	bool ScriptIsCoplanar( HSCRIPT hArea ) const;
-	bool ScriptIsContiguous( HSCRIPT hArea ) const;
-	float ScriptComputeAdjacentConnectionHeightChange( HSCRIPT hArea ) const;
-	void ScriptRemoveOrthogonalConnections( int dir );
-	HSCRIPT ScriptGetElevator( void ) { return ToHScript( (CBaseEntity*)GetElevator() ); }
-	void ScriptGetElevatorAreas( HSCRIPT hTable );
-	HSCRIPT ScriptGetDoor( void ) { return ToHScript( GetDoor() ); }
-	Vector ScriptComputeClosestPointInPortal( HSCRIPT to, int dir, const Vector &fromPos ) const;
-
 private:
 	friend class CTFNavMesh;
 
@@ -195,27 +153,11 @@ private:
 	// Raid mode -------------------------------------------------
 
 	float m_distanceToBombTarget;
-
-	EHANDLE m_hDoor;
-
-	HSCRIPT	m_hScriptInstance;
 };
 
-inline HSCRIPT ToHScript( CNavArea *pArea )
+inline CTFNavArea* ToTFNavArea( HSCRIPT hScript )
 {
-	CTFNavArea* pTerrorArea = ( CTFNavArea* )pArea;
-	return ( pTerrorArea ) ? pTerrorArea->GetScriptInstance() : NULL;
-}
-
-inline HSCRIPT ToHScript( CTFNavArea *pArea )
-{
-	return ( pArea ) ? pArea->GetScriptInstance() : NULL;
-}
-
-template <> ScriptClassDesc_t *GetScriptDesc<CTFNavArea>( CTFNavArea * );
-inline CTFNavArea *ToNavArea( HSCRIPT hScript )
-{
-	return ( IsValid( hScript ) ) ? (CTFNavArea *)g_pScriptVM->GetInstanceValue( hScript, GetScriptDescForClass(CTFNavArea) ) : NULL;
+	return static_cast< CTFNavArea* >( ToNavArea( hScript ) );
 }
 
 inline float CTFNavArea::GetTravelDistanceToBombTarget( void ) const
