+++ src/game/server/tf/tf_player.h
@@ -734,8 +734,6 @@ public:
 	bool				IsViewingCYOAPDA( void ) const { return m_bViewingCYOAPDA; }
 	bool				IsRegenerating( void ) const { return m_bRegenerating; }
 
-	HSCRIPT				ScriptGetActiveWeapon( void ) { return ToHScript( GetActiveTFWeapon() ); }
-
 	void				ScriptAddCond( int nCond );
 	void				ScriptAddCondEx( int nCond, float flDuration, HSCRIPT hProvider );
 	void				ScriptRemoveCond( int nCond );
@@ -809,16 +807,7 @@ public:
 		RemoveAllItems( bRemoveSuit );
 	}
 
-	Vector	ScriptWeapon_ShootPosition();
-	bool	ScriptWeapon_CanUse( HSCRIPT hWeapon );
-	void	ScriptWeapon_Equip( HSCRIPT hWeapon );
-	void	ScriptWeapon_Drop( HSCRIPT hWeapon );
-	void	ScriptWeapon_DropEx( HSCRIPT hWeapon, Vector vecTarget, Vector vecVelocity );
-	void	ScriptWeapon_Switch( HSCRIPT hWeapon );
-	void	ScriptWeapon_SetLast( HSCRIPT hWeapon );
-	HSCRIPT	ScriptGetLastWeapon();
 	void ScriptEquipWearableViewModel( HSCRIPT hWearableViewModel );
-	bool ScriptIsFakeClient() const { return this->IsFakeClient(); }
 	int ScriptGetBotType() const { return this->GetBotType(); }
 	bool ScriptIsBotOfType(int nType) const { return this->IsBotOfType(nType); }
 
@@ -1428,11 +1417,6 @@ public:
 	void SetGrapplingHookTarget( CBaseEntity *pTarget, bool bShouldBleed = false );
 	HSCRIPT ScriptGetGrapplingHookTarget() const { return ToHScript( m_hGrapplingHookTarget.Get() ); }
 	void ScriptSetGrapplingHookTarget( HSCRIPT pTarget, bool bShouldBleed ) { return SetGrapplingHookTarget( ToEnt( pTarget ), bShouldBleed ); }
-
-	void AddHudHideFlags(int flags) { m_Local.m_iHideHUD |= flags; }
-	void RemoveHudHideFlags(int flags) { m_Local.m_iHideHUD &= ~flags; }
-	void SetHudHideFlags(int flags) { m_Local.m_iHideHUD = flags; }
-	int GetHudHideFlags() { return m_Local.m_iHideHUD; }
 
 	bool IsUsingActionSlot() const { return m_bUsingActionSlot; }
